feats: 新增 user repo 的註解與修正錯誤
